Move all component files into one directory
